Fix CFLAGS in tox.ini
Without this fix gcc 4.9 (the default C compiler on Ubuntu 14.10) aborts
with the following error:

    cc1: error: unrecognised debug output level " -O0"

As far as I can tell this happens because "-g -O0" is passed as a single
argv[] entry to the C compiler.

The 'setenv' parameter in tox.ini is not shell code, it doesn't need to
be quoted (and spaces around '=' are also okay[*], so I added them, to make
it apparent this is not shell syntax).

[*] http://tox.readthedocs.org/en/latest/example/basic.html#setting-environment-variables