Extract magic string in ElementInclude._include()
ElementInclude defines several constants to refer to the relevant namespace - `XINCLUDE`, `XINCLUDE_INCLUDE` and `XINCLUDE_FALLBACK`. These constants are then used within `_include()`.

This change takes a remaining magic use of the namespace within `_include()` and extracts it to become a constant with the others. Doing this makes it possible to utilise an alternative namespace without having to modify `_include()`.