Fixing possible memory corruption if node is moved between docs
etree.insert function tries to handle the case when a node is moved
between documents with the function moveNodeToDocument. So far the
source_doc is taken from the destination node which is wrong.
The moveNodeToDocument function will not fix the names in the
document dictionaries because source and target doc are the same.

The fix takes now the source_doc from the node element which
should be inserted.

This fixes issue https://bugs.launchpad.net/lxml/+bug/1773749