Fix regression unknown error/undeclared variables while performing a xslt transform
Regression came up while updating from lxml 2.2.8. In a complex environment parameters were
suddenly reported as not declared like:
<xslt>:100:0:ERROR:XSLT:ERR_OK: runtime error, element 'result'
<string>:0:0:ERROR:XSLT:ERR_OK: Variable 'name' has not been declared.
<string>:0:0:ERROR:XSLT:ERR_OK: unknown error

Running the same transformation with the same template and inputs outside the
complex environment does not reproduce the error. The "complex environment" did several
xml operations before across multiple threads which seem to be important to reproduce the
issue.

At the end the issue seem to be caused by the string "name" which is contained in two dicts
and therefore with different addresses. The xsltXPathVariableLookup function does only a pointer
comparison and so "name" was != "name".