Would it be possible to support Python on iOS?
Both of the major implementations of Python 3 on iOS have long standing issues about supporting lxml. omz/Pythonista-Issues#245 and ColdGrub1384/Pyto#25  However, Apple rejects their attempts to add lxml support because the AppStore guidelines prohibit accessing private APIs.  Is there a way for lxml to support iOS as first class platform in such a way that lxml could be added without violating the AppStore guidelines?

Unrelated to the iOS questions above...  This PR proposes removing the __sudo:__ tag because it has become a no-op now that sudo is _always_ available on Travis CI.