Fix inheritance in lxml.html
As the old comment / FIXME from
8132c755adad4a75ba855d985dd257493bccc7fd notes, the mixin should come
first for the inheritance to be correct (the left-most class is the
first in the MRO, at least if no diamond inheritance is involved).

Also fix the odd `super` in `HtmlMixin` likely stemming from the
incorrect MRO.

Fixes the inheritance order of all `HTML*` base classes though it
probably doesn't matter for other than `HtmlElement`.