Validate that host_whitelist is not a string
Attacker can use `https:///evil.com` to make a malformed "hostless" URL
that would have a `netloc == ''` -- which is `in` any string. Strings
are not documented to be allowed in this config variable anyhow, so just
raise a type error if someone passes in a string by accident.

(This is a breaking change for people who didn't follow the documented
types, but shouldn't affect anyone else.)

New test fails on current master.