htmldiff: Avoid incorrectly identifying balanced/unbalanced tags in `merge_insert`
In the recently revised implementation of `merge_insert` we were checking
whether a tag exists in the `balanced/unbalanced` tags list by referring to
just the tag itself. This is unreliable since the same tag might exist in both
list. Incorrect identification leads to skipping of the `<ins>` tags in some
cases, resulting in incorrect diff being rendered.

This pathset fixes the issue described and adds a test to validate it.