Build: Fix Makefile to work on systems without `which(1)` (GH-416)
Replace the use of third-party `which(1)` executable with the shell
`command -v` builtin.  This avoids unnecessary dependency
on a third-party executable, in favor of a shell builtin that is
guaranteed by POSIX, and therefore fixes building documentation
on systems that no longer installs `which(1)` by default,
such as Gentoo Linux.