Return GenericAlias from __class_getattr__.
Make ElementTree a subclass of typing.Generic instead of adding
a __class_getattr__ method.

For the other scriptable classes, subclassing Generic conflicts with cython,
so we stick to __class_getattr__ returning GenericAlias.

For testing, check the return values of typing.get_origin and typing.get_args.
These functions seem to be the cannonical consumers of this data.