Allow subscripting generic types in annotations. (GH-461)
Some classes like lxml.etree.ElementTree are generic types that can contain different kinds of elements.
The type of these elements has to be specified in type annotations. For example:

element_tree: lxml.etree.ElementTree[lxml.etree.Element]

This requires adding a "__class_getitem__" class method to these class, as specified in PEP 560.
This commit adds these class methods.

ElementTree is now a subclass of "typing.Generic" instead of adding a __class_getattr__ method.
For the other scriptable classes, subclassing Generic conflicts with Cython,
so we stick to __class_getattr__ returning GenericAlias.

For testing, we check the return values of "typing.get_origin" and "typing.get_args".
These functions seem to be the canonical consumers of this data.

The list of generic types in lxml was taken from:

https://github.com/abelcheung/types-lxml/wiki/Using-specialised-class-directly